project/requirements.txt
pip>=25.3
setuptools>=68.0
wheel>=0.40

# Telegram бот
python-telegram-bot==20.5

# Для работы с .env (если будешь использовать переменные окружения)
python-dotenv==1.1.1
aiogram==3.1.0
# Для генерации случайных чисел и работы со временем (TTL кодов)
# Модули random и time входят в стандартную библиотеку, устанавливать их не нужно

# (Опционально) Для будущего хранения временных данных в Redis
redis==5.3.0

